requirements.txt
boto3==1.4.2
cmsplugin-filer==1.1.3
Django==1.8.14
django-appconf==1.0.2
django-ckeditor==5.2.1
django-ckeditor-filebrowser-filer==0.2.1
django-classy-tags==0.8.0
django-cms==3.2.5
django-constance[database]==2.0.0
djangocms-admin-style==1.2.4
djangocms-attributes-field==0.1.1
djangocms-column==1.6.0
djangocms-googlemap==0.5.2
djangocms-inherit==0.2.2
djangocms-installer==0.8.12
djangocms-link==1.8.3
djangocms-style==1.7.0
djangocms-text-ckeditor==2.9.5
django-environ==0.4.3
django-extensions==1.7.5
django-filer==1.2.5
django-formtools==1.0
django-mptt==0.8.6
django-polymorphic==0.8.1
django-reversion==1.8.7
django-s3direct==0.4.2
django-sekizai==0.10.0
Django-Select2==5.8.9
django-storages==1.5.1
django-treebeard==4.0.1
easy-thumbnails==2.3
gunicorn==19.5.0
html5lib==0.9999999
Pillow==2.3.2
psycopg2==2.6.2
pytz==2016.6.1
requests==2.12.4
six==1.10.0
tzlocal==1.2.2
Unidecode==0.4.19
whitenoise==3.1
wsgiref==0.1.2
